TrixiAtmo

Documentation for TrixiAtmo.

TrixiAtmo.TrixiAtmo

TrixiAtmo.MetricTermsCrossProduct

TrixiAtmo.MetricTermsCurlInvariant

TrixiAtmo.ShallowWaterEquations3D

Trixi.flux_fjordholm_etal

Trixi.flux_wintermeyer_etal

TrixiAtmo.P4estMeshCubedSphere2D

TrixiAtmo.examples_dir

TrixiAtmo.TrixiAtmo — Module

🌍 TrixiAtmo 🌍

TrixiAtmo.jl is a simulation package for atmospheric models based on Trixi.jl

See also: trixi-framework/TrixiAtmo.jl

TrixiAtmo.MetricTermsCrossProduct — Type

MetricTermsCrossProduct()

Struct used for multiple dispatch on functions that compute the metric terms. When the argument metric_terms is of type MetricTermsCrossProduct, the contravariant vectors are computed using the cross-product form.

TrixiAtmo.MetricTermsCurlInvariant — Type

MetricTermsCurlInvariant()

Struct used for multiple dispatch on functions that compute the metric terms. When the argument metric_terms is of type MetricTermsCurlInvariant, the contravariant vectors are computed using the curl invariant form.

TrixiAtmo.ShallowWaterEquations3D — Type

ShallowWaterEquations3D(; gravity, H0 = 0)

Shallow water equations (SWE) in three space dimensions in conservation form (with constant bottom topography). The equations are given by

\[\begin{aligned} \frac{\partial h}{\partial t} + \frac{\partial}{\partial x}(h v_1) + \frac{\partial}{\partial y}(h v_2) + \frac{\partial}{\partial z}(h v_3) &= 0 \\ \frac{\partial}{\partial t}(h v_1) + \frac{\partial}{\partial x}\left(h v_1^2 + \frac{g}{2}h^2\right) + \frac{\partial}{\partial y}(h v_1 v_2) + \frac{\partial}{\partial z}(h v_1 v_3) &= 0 \\ \frac{\partial}{\partial t}(h v_2) + \frac{\partial}{\partial x}(h v_1 v_2) + \frac{\partial}{\partial y}\left(h v_2^2 + \frac{g}{2}h^2\right) + \frac{\partial}{\partial z}(h v_2 v_3) &= 0 \\ \frac{\partial}{\partial t}(h v_3) + \frac{\partial}{\partial x}(h v_1 v_3) + \frac{\partial}{\partial y}(h v_2 v_3) + \frac{\partial}{\partial z}\left(h v_3^2 + \frac{g}{2}h^2\right) &= 0. \end{aligned}\]

The unknown quantities of the SWE are the water height $h$ and the velocities $\mathbf{v} = (v_1, v_2, v_3)^T$. The gravitational constant is denoted by g.

The additional quantity $H_0$ is also available to store a reference value for the total water height that is useful to set initial conditions or test the "lake-at-rest" well-balancedness.

In addition to the unknowns, Trixi.jl currently stores the bottom topography values at the approximation points despite being fixed in time. This is done for convenience of computing the bottom topography gradients on the fly during the approximation as well as computing auxiliary quantities like the total water height $H$ or the entropy variables. This affects the implementation and use of these equations in various ways:

The flux values corresponding to the bottom topography must be zero.

The bottom topography values must be included when defining initial conditions, boundary conditions or source terms.

AnalysisCallback analyzes this variable.

Trixi.jl's visualization tools will visualize the bottom topography by default.

References for the SWE are many but a good introduction is available in Chapter 13 of the book:

Randall J. LeVeque (2002) Finite Volume Methods for Hyperbolic Problems DOI: 10.1017/CBO9780511791253

Trixi.flux_fjordholm_etal — Method

flux_fjordholm_etal(u_ll, u_rr, orientation,
                    equations::ShallowWaterEquations1D)

Total energy conservative (mathematical entropy for shallow water equations). When the bottom topography is nonzero this should only be used as a surface flux otherwise the scheme will not be well-balanced. For well-balancedness in the volume flux use flux_wintermeyer_etal.

Details are available in Eq. (4.1) in the paper:

Ulrik S. Fjordholm, Siddhartha Mishr and Eitan Tadmor (2011) Well-balanced and energy stable schemes for the shallow water equations with discontinuous topography DOI: 10.1016/j.jcp.2011.03.042

Trixi.flux_wintermeyer_etal — Method

flux_wintermeyer_etal(u_ll, u_rr, orientation_or_normal_direction,
                      equations::ShallowWaterEquations2D)

Total energy conservative (mathematical entropy for shallow water equations) split form. When the bottom topography is nonzero this scheme will be well-balanced when used as a volume_flux. For the surface_flux either flux_wintermeyer_etal or flux_fjordholm_etal can be used to ensure well-balancedness and entropy conservation.

Further details are available in Theorem 1 of the paper:

Niklas Wintermeyer, Andrew R. Winters, Gregor J. Gassner and David A. Kopriva (2017) An entropy stable nodal discontinuous Galerkin method for the two dimensional shallow water equations on unstructured curvilinear meshes with discontinuous bathymetry DOI: 10.1016/j.jcp.2017.03.036

TrixiAtmo.P4estMeshCubedSphere2D — Method

P4estMeshCubedSphere2D(trees_per_face_dimension, radius;
                        polydeg, RealT=Float64,
                        initial_refinement_level=0, unsaved_changes=true,
                        p4est_partition_allow_for_coarsening=true)

Build a "Cubed Sphere" mesh as a 2D P4estMesh with 6 * trees_per_face_dimension^2 trees.

The mesh will have no boundaries.

Arguments

trees_per_face_dimension::Integer: the number of trees in the two local dimensions of each face.

radius::Integer: the radius of the sphere.

polydeg::Integer: polynomial degree used to store the geometry of the mesh. The mapping will be approximated by an interpolation polynomial of the specified degree for each tree.

RealT::Type: the type that should be used for coordinates.

initial_refinement_level::Integer: refine the mesh uniformly to this level before the simulation starts.

unsaved_changes::Bool: if set to true, the mesh will be saved to a mesh file.

p4est_partition_allow_for_coarsening::Bool: Must be true when using AMR to make mesh adaptivity independent of domain partitioning. Should be false for static meshes to permit more fine-grained partitioning.

TrixiAtmo.examples_dir — Method

examples_dir()

Return the directory where the example files provided with TrixiAtmo.jl are located. If TrixiAtmo.jl is installed as a regular package (with ]add Trixi), these files are read-only and should not be modified. To find out which files are available, use, e.g., readdir:

Examples

readdir(examples_dir())