TrixiAtmo

Documentation for TrixiAtmo.

TrixiAtmo.TrixiAtmo

TrixiAtmo.AbstractCovariantEquations

TrixiAtmo.CovariantLinearAdvectionEquation2D

TrixiAtmo.MetricTermsCrossProduct

TrixiAtmo.MetricTermsInvariantCurl

TrixiAtmo.ShallowWaterEquations3D

Trixi.flux_fjordholm_etal

Trixi.flux_nonconservative_fjordholm_etal

Trixi.flux_nonconservative_wintermeyer_etal

Trixi.flux_wintermeyer_etal

TrixiAtmo.P4estMeshCubedSphere2D

TrixiAtmo.P4estMeshQuadIcosahedron2D

TrixiAtmo.clean_solution_lagrange_multiplier!

TrixiAtmo.examples_dir

TrixiAtmo.have_aux_node_vars

TrixiAtmo.initial_condition_gaussian

TrixiAtmo.source_terms_lagrange_multiplier

TrixiAtmo.transform_to_cartesian

TrixiAtmo.transform_to_contravariant

TrixiAtmo.TrixiAtmo — Module

🌍 TrixiAtmo 🌍

TrixiAtmo.jl is a simulation package for atmospheric models based on Trixi.jl

See also: trixi-framework/TrixiAtmo.jl

TrixiAtmo.AbstractCovariantEquations — Type

AbstractCovariantEquations{NDIMS, 
                           NDIMS_AMBIENT, 
                           NVARS} <: AbstractEquations{NDIMS, NVARS}

Abstract type used to dispatch on systems of equations in covariant form, in which fluxes and prognostic variables are stored and computed in terms of their contravariant components defining their expansions in terms of the local covariant tangent basis. The type parameter NDIMS denotes the dimension of the manifold on which the equations are solved, while NDIMS_AMBIENT is the dimension of the ambient space in which such a manifold is embedded. Some references on discontinuous Galerkin methods in covariant flux form are listed below:

M. Baldauf (2020). Discontinuous Galerkin solver for the shallow-water equations in covariant form on the sphere and the ellipsoid. Journal of Computational Physics 410:109384. DOI: 10.1016/j.jcp.2020.109384

M. Baldauf (2021). A horizontally explicit, vertically implicit (HEVI) discontinuous Galerkin scheme for the 2-dimensional Euler and Navier-Stokes equations using terrain-following coordinates. Journal of Computational Physics 446:110635. DOI: 10.1016/ j.jcp.2021.110635

L. Bao, R. D. Nair, and H. M. Tufo (2014). A mass and momentum flux-form high-order discontinuous Galerkin shallow water model on the cubed-sphere. A mass and momentum flux-form high-order discontinuous Galerkin shallow water model on the cubed-sphere. Journal of Computational Physics 271:224-243. DOI: 10.1016/j.jcp.2013.11.033

When using this equation type, functions which are evaluated pointwise, such as fluxes, source terms, and initial conditions take in the extra argument aux_vars, which contains the geometric information needed for the covariant form. To convert an initial condition given in terms of global spherical velocity or momentum components to one given in terms of local contravariant components, see transform_to_contravariant.

TrixiAtmo.CovariantLinearAdvectionEquation2D — Type

CovariantLinearAdvectionEquation2D <: AbstractCovariantEquations{2, 3, 3}

A variable-coefficient linear advection equation can be defined on a two-dimensional manifold $S \subset \mathbb{R}^3$ as

\[\partial_t h + \nabla_S \cdot (h \vec{v}) = 0,\]

where $\nabla_S \cdot$ is the horizontal divergence operator on $S$. We treat this problem as a system of equations in which the first variable is the scalar conserved quantity $h$, and the second two are the contravariant components $v^1$ and $v^2$ used in the expansion with respect to the covariant basis vectors $\vec{a}_1$ and $\vec{a}_2$ as

\[\vec{v} = v^1 \vec{a}_1 + v^2 \vec{a}_2.\]

where $\vec{a}_1 = \partial \vec{x} / \partial \xi^1$ and $\vec{a}_2 = \partial \vec{x} / \partial \xi^2$ are the so-called covariant basis vectors, and $\xi^1$ and $\xi^2$ are the local reference space coordinates. The velocity components are spatially varying but assumed to be constant in time, so we do not apply any flux or dissipation to such variables. The resulting system is then given on the reference element as

\[\sqrt{G} \frac{\partial}{\partial t}\left[\begin{array}{c} h \\ v^1 \\ v^2 \end{array}\right] + \frac{\partial}{\partial \xi^1} \left[\begin{array}{c} \sqrt{G} h v^1 \\ 0 \\ 0 \end{array}\right] + \frac{\partial}{\partial \xi^2} \left[\begin{array}{c} \sqrt{G} h v^2 \\ 0 \\ 0 \end{array}\right] = \left[\begin{array}{c} 0 \\ 0 \\ 0 \end{array}\right],\]

where $G$ is the determinant of the covariant metric tensor expressed as a matrix with entries $G_{ij} = \vec{a}_i \cdot \vec{a}_j$. Note that the variable advection velocity components could also be stored as auxiliary variables, similarly to the geometric information.

TrixiAtmo.MetricTermsCrossProduct — Type

MetricTermsCrossProduct()

Struct used for multiple dispatch on functions that compute the metric terms. When the argument metric_terms is of type MetricTermsCrossProduct, the contravariant vectors are computed using the cross-product form.

TrixiAtmo.MetricTermsInvariantCurl — Type

MetricTermsInvariantCurl()

Struct used for multiple dispatch on functions that compute the metric terms. When the argument metric_terms is of type MetricTermsInvariantCurl, the contravariant vectors are computed using the invariant curl form.

References

Kopriva, D. A. (2006). Metric identities and the discontinuous spectral element method on curvilinear meshes. Journal of Scientific Computing 26, 301-327. DOI: 10.1007/s10915-005-9070-8

Vinokur, M. and Yee, H. C. (2001). Extension of efficient low dissipation high order schemes for 3-D curvilinear moving grids. In Caughey, D. A., and Hafez, M. M. (eds.), Frontiers of Computational Fluid Dynamics 2002, World Scientific, Singapore, pp. 129–164. DOI: 10.1142/9789812810793_0008

TrixiAtmo.ShallowWaterEquations3D — Type

ShallowWaterEquations3D(; gravity, H0 = 0)

Shallow water equations (SWE) in three space dimensions in conservation form (with constant bottom topography). The equations are given by

\[\begin{aligned} \frac{\partial h}{\partial t} + \frac{\partial}{\partial x}(h v_1) + \frac{\partial}{\partial y}(h v_2) + \frac{\partial}{\partial z}(h v_3) &= 0 \\ \frac{\partial}{\partial t}(h v_1) + \frac{\partial}{\partial x}\left(h v_1^2 + \frac{g}{2}h^2\right) + \frac{\partial}{\partial y}(h v_1 v_2) + \frac{\partial}{\partial z}(h v_1 v_3) &= 0 \\ \frac{\partial}{\partial t}(h v_2) + \frac{\partial}{\partial x}(h v_1 v_2) + \frac{\partial}{\partial y}\left(h v_2^2 + \frac{g}{2}h^2\right) + \frac{\partial}{\partial z}(h v_2 v_3) &= 0 \\ \frac{\partial}{\partial t}(h v_3) + \frac{\partial}{\partial x}(h v_1 v_3) + \frac{\partial}{\partial y}(h v_2 v_3) + \frac{\partial}{\partial z}\left(h v_3^2 + \frac{g}{2}h^2\right) &= 0. \end{aligned}\]

The unknown quantities of the SWE are the water height $h$ and the velocities $\mathbf{v} = (v_1, v_2, v_3)^T$. The gravitational constant is denoted by g.

The 3D Shallow Water Equations (SWE) extend the 2D SWE to model shallow water flows on 2D manifolds embedded within 3D space. To confine the flow to the 2D manifold, a source term incorporating a Lagrange multiplier is applied. This term effectively removes momentum components that are normal to the manifold, ensuring the flow remains constrained within the 2D surface.

The additional quantity $H_0$ is also available to store a reference value for the total water height that is useful to set initial conditions or test the "lake-at-rest" well-balancedness.

In addition to the unknowns, Trixi.jl currently stores the bottom topography values at the approximation points despite being fixed in time. This is done for convenience of computing the bottom topography gradients on the fly during the approximation as well as computing auxiliary quantities like the total water height $H$ or the entropy variables. This affects the implementation and use of these equations in various ways:

The flux values corresponding to the bottom topography must be zero.

The bottom topography values must be included when defining initial conditions, boundary conditions or source terms.

AnalysisCallback analyzes this variable.

Trixi.jl's visualization tools will visualize the bottom topography by default.

References:

J. Coté (1988). "A Lagrange multiplier approach for the metric terms of semi-Lagrangian models on the sphere". Quarterly Journal of the Royal Meteorological Society 114, 1347-1352. DOI: 10.1002/qj.49711448310

F. X. Giraldo (2001). "A spectral element shallow water model on spherical geodesic grids". DOI: 10.1002/1097-0363(20010430)35:8<869::AID-FLD116>3.0.CO;2-S

Trixi.flux_fjordholm_etal — Method

flux_fjordholm_etal(u_ll, u_rr, orientation,
                    equations::ShallowWaterEquations1D)

Total energy conservative (mathematical entropy for shallow water equations). When the bottom topography is nonzero this should only be used as a surface flux otherwise the scheme will not be well-balanced. For well-balancedness in the volume flux use flux_wintermeyer_etal.

Details are available in Eq. (4.1) in the paper:

Ulrik S. Fjordholm, Siddhartha Mishr and Eitan Tadmor (2011) Well-balanced and energy stable schemes for the shallow water equations with discontinuous topography DOI: 10.1016/j.jcp.2011.03.042

Trixi.flux_nonconservative_fjordholm_etal — Method

flux_nonconservative_fjordholm_etal(u_ll, u_rr, orientation::Integer,
                                    equations::ShallowWaterEquations2D)
flux_nonconservative_fjordholm_etal(u_ll, u_rr,
                                    normal_direction::AbstractVector,
                                    equations::ShallowWaterEquations2D)

Non-symmetric two-point surface flux discretizing the nonconservative (source) term of that contains the gradient of the bottom topography ShallowWaterEquations2D.

This flux can be used together with flux_fjordholm_etal at interfaces to ensure entropy conservation and well-balancedness.

Further details for the original finite volume formulation are available in

Ulrik S. Fjordholm, Siddhartha Mishra and Eitan Tadmor (2011) Well-balanced and energy stable schemes for the shallow water equations with discontinuous topography DOI: 10.1016/j.jcp.2011.03.042

and for curvilinear 2D case in the paper:

Niklas Wintermeyer, Andrew R. Winters, Gregor J. Gassner and David A. Kopriva (2017) An entropy stable nodal discontinuous Galerkin method for the two dimensional shallow water equations on unstructured curvilinear meshes with discontinuous bathymetry DOI: 10.1016/j.jcp.2017.03.036

Trixi.flux_nonconservative_wintermeyer_etal — Method

flux_nonconservative_wintermeyer_etal(u_ll, u_rr, orientation::Integer,
                                      equations::ShallowWaterEquations2D)
flux_nonconservative_wintermeyer_etal(u_ll, u_rr,
                                      normal_direction::AbstractVector,
                                      equations::ShallowWaterEquations2D)

Non-symmetric two-point volume flux discretizing the nonconservative (source) term that contains the gradient of the bottom topography ShallowWaterEquations2D.

For the surface_flux either flux_wintermeyer_etal or flux_fjordholm_etal can be used to ensure well-balancedness and entropy conservation.

Further details are available in the papers:

Niklas Wintermeyer, Andrew R. Winters, Gregor J. Gassner and David A. Kopriva (2017) An entropy stable nodal discontinuous Galerkin method for the two dimensional shallow water equations on unstructured curvilinear meshes with discontinuous bathymetry DOI: 10.1016/j.jcp.2017.03.036

Patrick Ersing, Andrew R. Winters (2023) An entropy stable discontinuous Galerkin method for the two-layer shallow water equations on curvilinear meshes DOI: 10.48550/arXiv.2306.12699

Trixi.flux_wintermeyer_etal — Method

flux_wintermeyer_etal(u_ll, u_rr, orientation_or_normal_direction,
                      equations::ShallowWaterEquations2D)

Total energy conservative (mathematical entropy for shallow water equations) split form. When the bottom topography is nonzero this scheme will be well-balanced when used as a volume_flux. For the surface_flux either flux_wintermeyer_etal or flux_fjordholm_etal can be used to ensure well-balancedness and entropy conservation.

Further details are available in Theorem 1 of the paper:

Niklas Wintermeyer, Andrew R. Winters, Gregor J. Gassner and David A. Kopriva (2017) An entropy stable nodal discontinuous Galerkin method for the two dimensional shallow water equations on unstructured curvilinear meshes with discontinuous bathymetry DOI: 10.1016/j.jcp.2017.03.036

TrixiAtmo.P4estMeshCubedSphere2D — Method

P4estMeshCubedSphere2D(trees_per_face_dimension, radius;
                        polydeg, RealT=Float64,
                        initial_refinement_level=0, unsaved_changes=true,
                        p4est_partition_allow_for_coarsening=true,
                        element_local_mapping=false)

Build a "Cubed Sphere" mesh as a 2D P4estMesh with 6 * trees_per_face_dimension^2 trees.

The mesh will have no boundaries.

Arguments

trees_per_face_dimension::Integer: the number of trees in the two local dimensions of each face.

radius::Integer: the radius of the sphere.

polydeg::Integer: polynomial degree used to store the geometry of the mesh. The mapping will be approximated by an interpolation polynomial of the specified degree for each tree.

RealT::Type: the type that should be used for coordinates.

initial_refinement_level::Integer: refine the mesh uniformly to this level before the simulation starts.

unsaved_changes::Bool: if set to true, the mesh will be saved to a mesh file.

p4est_partition_allow_for_coarsening::Bool: Must be true when using AMR to make mesh adaptivity independent of domain partitioning. Should be false for static meshes to permit more fine-grained partitioning.

element_local_mapping::Bool: option to use the alternative element-local mapping from Appendix A of Guba et al. (2014). If set to true, the four corner vertex positions for each element will be obtained through an equiangular gnomonic projection (Ronchi et al. 1996), and the tree node coordinates within the element (i.e. the field tree_node_coordinates) will be obtained by first using a bilinear mapping based on the four corner vertices, and then projecting the bilinearly mapped nodes onto the spherical surface by normalizing the resulting Cartesian coordinates and scaling by radius. If set to false, the equiangular gnomonic projection will be used for all tree node coordinates.

Warning

Adaptivity and MPI parallelization are not yet supported for equations in covariant form, and we require initial_refinement_level = 0 for such cases. Furthermore, the calculation of the metric terms for the covariant form currently requires polydeg to be equal to the polynomial degree of the solver, and element_local_mapping = true.

!!!

TrixiAtmo.P4estMeshQuadIcosahedron2D — Method

P4estMeshQuadIcosahedron2D(trees_per_face_dimension, radius;
                           polydeg, RealT=Float64,
                           initial_refinement_level=0, unsaved_changes=true,
                           p4est_partition_allow_for_coarsening=true)

Build a quad-based icosahedral mesh as a 2D P4estMesh with 60 * trees_per_face_dimension^2 trees (20 triangular faces of the icosahedron, each subdivided into 3 parent quads, each of which subdivided into trees_per_face_dimension^2 trees).

The node coordinates of the trees will be obtained using the element-local mapping from Appendix A of Guba et al. (2014). See P4estMeshCubedSphere2D for more information about the element-local mapping.

The mesh will have no boundaries.

Arguments

trees_per_face_dimension::Integer: the number of trees in the two local dimensions of each parent quad.

radius::Integer: the radius of the sphere.

polydeg::Integer: polynomial degree used to store the geometry of the mesh. The mapping will be approximated by an interpolation polynomial of the specified degree for each tree.

RealT::Type: the type that should be used for coordinates.

initial_refinement_level::Integer: refine the mesh uniformly to this level before the simulation starts.

unsaved_changes::Bool: if set to true, the mesh will be saved to a mesh file.

p4est_partition_allow_for_coarsening::Bool: Must be true when using AMR to make mesh adaptivity independent of domain partitioning. Should be false for static meshes to permit more fine-grained partitioning.

Warning

Adaptivity and MPI parallelization are not yet supported for equations in covariant form, and we require initial_refinement_level = 0 for such cases. Furthermore, the calculation of the metric terms for the covariant form currently requires polydeg to be equal to the polynomial degree of the solver.

!!!

TrixiAtmo.clean_solution_lagrange_multiplier! — Method

     clean_solution_lagrange_multiplier!(u, equations::ShallowWaterEquations3D, normal_direction)

Function to apply Lagrange multiplier discretely to the solution in order to constrain the momentum to a 2D manifold.

The vector normal_direction is perpendicular to the 2D manifold. By default, this is the normal contravariant basis vector.

TrixiAtmo.examples_dir — Method

examples_dir()

Return the directory where the example files provided with TrixiAtmo.jl are located. If TrixiAtmo.jl is installed as a regular package (with ]add Trixi), these files are read-only and should not be modified. To find out which files are available, use, e.g., readdir:

Examples

readdir(examples_dir())

TrixiAtmo.have_aux_node_vars — Method

have_aux_node_vars(equations)

Trait function determining whether equations requires the use of auxiliary variables. Classical conservation laws such as the CompressibleEulerEquations2D do not require auxiliary variables. The return value will be True() or False() to allow dispatching on the return type.

TrixiAtmo.initial_condition_gaussian — Method

initial_condition_gaussian(x, t)

This Gaussian bell case is a smooth initial condition suitable for testing the convergence of discretizations of the linear advection equation on a spherical domain of radius $a = 6. 37122 \times 10^3\ \mathrm{m}$, representing the surface of the Earth. The height field is given in terms of the Cartesian coordinates $x$, $y$, $z$ as

\[h(x,y,z) = h_0 \exp\Big(-b_0 \big((x-x_0)^2 + (y-y_0)^2 + (z-z_0)^2\big) \Big),\]

where $h_0 = 1 \times 10^3\ \mathrm{m}$ is the height of the bell, $b_0 = 5 / a$ is the width parameter, and the Cartesian coordinates of the bell's centre at longitude $\lambda_0 = 3\pi/2$ and latitude $\theta_0 = 0$ are given by

\[\begin{aligned} x_0 &= a\cos\theta_0\cos\lambda_0,\\ y_0 &= a\cos\theta_0\sin\lambda_0,\\ z_0 &= a\sin\theta_0. \end{aligned}\]

The velocity field corresponds to a solid body rotation with a period of 12 days, at an angle of $\alpha = \pi/4$ from the polar axis. The zonal and meridional components of the velocity field are then given in terms of the longitude $\lambda$ and latitude $\theta$ as

\[\begin{aligned} u(\lambda,\theta) &= V (\cos\theta\cos\alpha + \sin\theta\cos\lambda\sin\alpha),\\ v(\lambda,\theta) &= -V \sin\lambda\sin\alpha, \end{aligned}\]

where we take $V = 2\pi a / 12\ \mathrm{days}$. This test case is adapted from Case 1 of the test suite described in the following paper:

D. L. Williamson, J. B. Drake, J. J. Hack, R. Jakob, and P. N. Swarztrauber (1992). A standard test set for numerical approximations to the shallow water equations in spherical geometry. Journal of Computational Physics, 102(1):211-224. DOI: 10.1016/S0021-9991(05)80016-6

This function returns SVector(h, vlon, vlat, b), where the first three entries are the height, zonal velocity, and meridional velocity. The fourth entry, representing variable bottom topography, is set to zero. The functions transform_to_contravariant and transform_to_cartesian are available for converting to the prognostic variables for Cartesian and covariant formulations.

TrixiAtmo.source_terms_lagrange_multiplier — Method

source_terms_lagrange_multiplier(u, du, x, t,
                                      equations::ShallowWaterEquations3D,
                                      normal_direction)

Source term function to apply a Lagrange multiplier to the semi-discretization in order to constrain the momentum to a 2D manifold.

The vector normal_direction is perpendicular to the 2D manifold. By default, this is the normal contravariant basis vector.

TrixiAtmo.transform_to_cartesian — Method

transform_to_cartesian(initial_condition, equations)

Takes in a function with the signature initial_condition(x, t) which returns an initial condition given in terms of zonal and meridional velocity or momentum components, and returns another function with the signature initial_condition_transformed(x, t, equations) which returns the same initial condition with the velocity or momentum vector given in terms of Cartesian components.

TrixiAtmo.transform_to_contravariant — Method

transform_to_contravariant(initial_condition, equations)

Takes in a function with the signature initial_condition(x, t) which returns an initial condition given in terms of zonal and meridional velocity or momentum components, and returns another function with the signature initial_condition_transformed(x, t, aux_vars, equations) which returns the same initial condition with the velocity or momentum vector given in terms of contravariant components.