TrixiAtmo

Documentation for TrixiAtmo.

TrixiAtmo.TrixiAtmo

TrixiAtmo.AbstractCovariantEquations

TrixiAtmo.CovariantLinearAdvectionEquation2D

TrixiAtmo.P4estElementContainerCovariant

TrixiAtmo.P4estMeshCubedSphere2D

TrixiAtmo.examples_dir

TrixiAtmo.TrixiAtmo — Module

🌍 TrixiAtmo 🌍

TrixiAtmo.jl is a simulation package for atmospheric models based on Trixi.jl

See also: trixi-framework/TrixiAtmo.jl

TrixiAtmo.AbstractCovariantEquations — Type

AbstractCovariantEquations{NDIMS, 
                           NDIMS_AMBIENT, 
                           NVARS} <: AbstractEquations{NDIMS, NVARS}

Abstract type used to dispatch on systems of equations in covariant form, in which fluxes and prognostic variables are stored and computed in terms of their contravariant components defining their expansions in terms of the local covariant tangent basis. The type parameter NDIMS denotes the dimension of the manifold on which the equations are solved, while NDIMS_AMBIENT is the dimension of the ambient space in which such a manifold is embedded.

Note

Components of vector-valued fields should be prescibed within the global coordinate system (i.e. zonal and meridional components in the case of a spherical shell). When dispatched on this type, the function Trixi.compute_coefficients! will internally use the covariant_basis field of the container type P4estElementContainerCovariant to obtain the local contravariant components used in the solver.

TrixiAtmo.CovariantLinearAdvectionEquation2D — Type

CovariantLinearAdvectionEquation2D <: AbstractCovariantEquations{2, 3, 3}

A variable-coefficient linear advection equation can be defined on a two-dimensional manifold $S \subset \mathbb{R}^3$ as

\[\partial_t h + \nabla_S \cdot (h \vec{v}) = 0.\]

We treat this problem as a system of equations in which the first variable is the scalar conserved quantity $h$, and the second two are the contravariant components $v^1$ and $v^2$ used in the expansion

\[\vec{v} = v^1 \vec{a}_1 + v^2 \vec{a}_2,\]

which are spatially varying but remain constant in time (i.e. no flux or dissipation is applied to such variables). The resulting system is then given on the reference element as

\[J \frac{\partial}{\partial t}\left[\begin{array}{c} h \\ v^1 \\ v^2 \end{array}\right] + \frac{\partial}{\partial \xi^1} \left[\begin{array}{c} J h v^1 \\ 0 \\ 0 \end{array}\right] + \frac{\partial}{\partial \xi^2} \left[\begin{array}{c} J h v^2 \\ 0 \\ 0 \end{array}\right] = \left[\begin{array}{c} 0 \\ 0 \\ 0 \end{array}\right],\]

where $J$ is the area element (see the documentation for P4estElementContainerCovariant).

Note

The initial condition should be prescribed as $[h, u, v]^{\mathrm{T}}$ in terms of the global velocity components $u$ and $v$ (i.e. zonal and meridional components in the case of a spherical shell). The transformation to local contravariant components $v^1$ and $v^2$ is handled internally within Trixi.compute_coefficients!.

TrixiAtmo.P4estElementContainerCovariant — Type

P4estElementContainerCovariant{NDIMS, RealT <: Real, uEltype <: Real,
                               NDIMSP1, NDIMSP2,
                               NDIMSP3, NDIMSP4} <: Trixi.AbstractContainer

Specialized element container for equations in covariant form on a manifold of dimension NDIMS, created when Trixi.init_elements is dispatched on AbstractCovariantEquations{NDIMS, NDIMS_AMBIENT} and P4estMesh{NDIMS, NDIMS_AMBIENT. Contains the following geometric information:

node_coordinates::Array{RealT, NDIMSP2}: Cartesian components of the node positions within the ambient space of dimension NDIMS_AMBIENT, where the first index is the Cartesian component index, the next NDIMS indices are tensor-product node indices ($i$, $j$, for NDIMS = 2 or $i$, $j$, $k$, for NDIMS = 3), and the last is the element index.

covariant_basis::Array{RealT, NDIMSP3}: Components of the covariant tangent basis vectors $\vec{a}_m = \partial \vec{X} / \partial \xi^m$ expanded in terms of a global tangent basis (i.e. zonal and meridional components in the case of a spherical shell), where $\vec{X}$ denotes the mapping from the reference element to the physical element. he first index is the global component index $l$, the second index is the local component index $m$, the next NDIMS indices are tensor-product node indices ($i$, $j$, for NDIMS = 2 or $i$, $j$, $k$, for NDIMS = 3), and the last is the element index.

inverse_jacobian::Array{RealT, NDIMSP1}: Nodal values of $1/J$, where $J = \sqrt{\mathrm{det}(\bm{G})}$, and $\bm{G}$ is the matrix of covariant metric tensor components $G_{lm} = \vec{a}_l \cdot \vec{a}_m$. The first NDIMS indices are tensor-product node indices ($i$, $j$, for NDIMS = 2 or $i$, $j$, $k$, for NDIMS = 3), and the last is the element index.

Warning

The covariant solver currently only supports conforming meshes.

TrixiAtmo.P4estMeshCubedSphere2D — Method

P4estMeshCubedSphere2D(trees_per_face_dimension, radius;
                        polydeg, RealT=Float64,
                        initial_refinement_level=0, unsaved_changes=true,
                        p4est_partition_allow_for_coarsening=true,
                        element_local_mapping=false)

Build a "Cubed Sphere" mesh as a 2D P4estMesh with 6 * trees_per_face_dimension^2 trees.

The mesh will have no boundaries.

Arguments

trees_per_face_dimension::Integer: the number of trees in the two local dimensions of each face.

radius::Integer: the radius of the sphere.

polydeg::Integer: polynomial degree used to store the geometry of the mesh. The mapping will be approximated by an interpolation polynomial of the specified degree for each tree.

RealT::Type: the type that should be used for coordinates.

initial_refinement_level::Integer: refine the mesh uniformly to this level before the simulation starts.

unsaved_changes::Bool: if set to true, the mesh will be saved to a mesh file.

p4est_partition_allow_for_coarsening::Bool: Must be true when using AMR to make mesh adaptivity independent of domain partitioning. Should be false for static meshes to permit more fine-grained partitioning.

element_local_mapping::Bool: option to use the element-local mapping from Appendix A of Guba et al. (2014).

TrixiAtmo.examples_dir — Method

examples_dir()

Return the directory where the example files provided with TrixiAtmo.jl are located. If TrixiAtmo.jl is installed as a regular package (with ]add Trixi), these files are read-only and should not be modified. To find out which files are available, use, e.g., readdir:

Examples

readdir(examples_dir())